feat: implement localization support for Chinese
