+++ lib/l10n/app_zh.arb
@@ -1,3 +1,3 @@
 {
-  "trayDashboardLabel": "Open Dashboard"
+  "trayDashboardLabel": "你好"
 }
