+++ lib/main.dart
@@ -2,17 +2,14 @@ import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:lux/app.dart';
 import 'package:lux/const/const.dart';
 import 'package:lux/core_config.dart';
 import 'package:lux/elevate.dart';
-import 'package:lux/home.dart';
 import 'package:lux/notifier.dart';
-import 'package:lux/tr.dart';
 import 'package:lux/utils.dart';
 import 'package:path/path.dart' as path;
 import 'package:window_manager/window_manager.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 void main(args) async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -48,26 +45,7 @@ void main(args) async {
     var isDarkMode = await readTheme() == ThemeType.dark;
     var theme = isDarkMode ? "dark" : "light";
     debugPrint("using theme: $theme");
-    runApp(MaterialApp(
-        theme: ThemeData(
-            scaffoldBackgroundColor: isDarkMode ? Color(darkBackgroundColor) : Colors.white), //Dark mode of dashboard
-        home: Home(theme),
-        onGenerateTitle: (context) {
-          initTr(context);
-          return 'Lux';
-        },
-      localizationsDelegates: [
-        AppLocalizations.delegate,
-        GlobalMaterialLocalizations.delegate,
-        GlobalWidgetsLocalizations.delegate,
-        GlobalCupertinoLocalizations.delegate,
-      ],
-      supportedLocales: [
-        Locale('en'),
-        Locale('zh'),
-      ],
-    )
-    );
+    runApp(App(theme, isDarkMode ? Color(darkBackgroundColor) : Colors.white));
   } catch (e) {
     await notifier.show("$e");
     exitApp();
@@ -75,13 +53,3 @@ void main(args) async {
 }
 
 
-class LocaleModel extends ChangeNotifier {
-  Locale? _locale;
-
-  Locale? get locale => _locale;
-
-  void set(Locale locale) {
-    _locale = locale;
-    notifyListeners();
-  }
-}
