+++ pubspec.yaml
@@ -59,6 +59,7 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   intl: any
+  provider: ^6.1.4
 
 
 dev_dependencies:
